+++ cmd/src/main.go
@@ -34,15 +34,17 @@ The commands are:
 
 	api             interacts with the Sourcegraph GraphQL API
 	batch           manages batch changes
+	code-intel      manages code intelligence data
 	config          manages global, org, and user settings
 	extensions,ext  manages extensions (experimental)
 	extsvc          manages external services
 	login           authenticate to a Sourcegraph instance with your user credentials
-	lsif            manages LSIF data
+	lsif            manages LSIF data (deprecated: use 'code-intel')
 	orgs,org        manages organizations
 	repos,repo      manages repositories
 	search          search for results on Sourcegraph
 	serve-git       serves your local git repositories over HTTP for Sourcegraph to pull
+	upload          upload an index to a Sourcegraph instance
 	users,user      manages users
 	version         display and compare the src-cli version against the recommended version for your instance